fineltest

+++ app/src/main/java/com/xdev/deliverytn/recyclerview/RecyclerViewOrderAdapter.java
@@ -89,7 +89,7 @@ public class RecyclerViewOrderAdapter extends RecyclerView.Adapter<OrderViewHold
             // populate the current row on the RecyclerView
             status = String.valueOf(order.status.charAt(0));
             price = Integer.toString(order.max_range);
-            charge = Integer.toString(order.deliveryCharge);
+            charge = Float.toString(order.deliveryCharge);
             holder.category.setText(order.category);
             setExpiry(order.expiryDate, order.expiryTime);
             String date = DateFormat.getDateInstance(DateFormat.MEDIUM).format(calendar.getTime());
